Agregando permisos a las rutas, validando el tipo de usuario que realiza las acciones.

+++ app/Livewire/Admin/Tags/DeleteTag.php
@@ -3,31 +3,49 @@
 namespace App\Livewire\Admin\Tags;
 
 use App\Models\Tag;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 use Livewire\Attributes\On;
 
 class DeleteTag extends Component
 {
+	use AuthorizesRequests;
+
 	public $tagId;
 	public $isModalOpen = false;
 
 	#[On('openModalDelete')]
 	public function openModalDelete($tagId)
 	{
-		$this->tagId = $tagId;
+		$this->authorize('tags.destroy');
+
+		$tag = Tag::findOrFail($tagId);
+		$this->tagId = $tag->id;
 		$this->isModalOpen = true;
 	}
 
-	#[On('closeDeleteTagModal')]
-	public function closeDeleteTagModal()
+	#[On('closeModalDelete')]
+	public function closeModalDelete()
 	{
 		$this->isModalOpen = false;
+		$this->reset('tagId');
 	}
 
 	public function deleteTag()
 	{
+		$this->authorize('tags.destroy');
+
 		try {
-			$tag = Tag::find($this->tagId);
+			$tag = Tag::findOrFail($this->tagId);
+
+			if($tag->posts()->count() > 0){
+				$this->dispatch('showAlert',
+					message: 'No se puede eliminar la etiqueta porque tiene posts asociados.',
+					type: 'error'
+				);
+				$this->closeModalDelete();
+				return;
+			}
 
 			if ($tag) {
 				$tag->delete();
@@ -37,12 +55,12 @@ class DeleteTag extends Component
 				$message = 'La etiqueta no existe o ya fue eliminada.';
 			}
 
-			$this->closeDeleteTagModal();
+			$this->closeModalDelete();
 			$this->dispatch('showAlert', message: $message);
 
 		} catch (\Exception $e) {
 
-			$this->closeDeleteTagModal();
+			$this->closeModalDelete();
 			$message = 'Ocurrió un error al eliminar la etiqueta. Por favor, inténtalo de nuevo.';
 			$this->dispatch('showAlert', message: $message);
 		}
